Namespace APIs and update in the controller
